refactor(shikomi-cli): 工程5 ペガサス指摘 — lib.rs 858行 → 338行 + record_runners.rs 416行
工程5 ペガサス指摘 (lib.rs 500 行ルール超過、858 行) 解消:

- crates/shikomi-cli/src/record_runners.rs (新規 416 行) に切り出し:
  - run_list / run_add / run_edit / run_remove (レコード CRUD dispatcher)
  - lookup_label_for_remove (run_remove 専用ヘルパ)
  - resolve_secret_value / read_value_from_stdin (値取得ヘルパ)
  - is_color_enabled (run_list 専用)
  - decide_kind_for_input / discriminant (Issue #33 fail-secure 純粋関数)
  - tests (TC-UT-130〜134 + discriminant 射影確認、Issue #33)

- crates/shikomi-cli/src/lib.rs (858 → 338 行、60% 削減) は以下に責務集約:
  - panic_hook + LOCALE_CACHE (secret 漏洩経路遮断)
  - RepositoryHandle / RepositoryHandleDiscriminant (enum 定義のみ)
  - run() コンポジションルート + run_vault dispatcher
  - build_handle / handle_clap_error / init_tracing / emit_error_and_exit
  - print_stdout / eprint_stderr (record_runners と build_handle 両方から
    使うため pub(crate) で公開)

修正:
- run() 内の `run_*` 直接呼出を `record_runners::run_*` 経由に置換
- 不要になった use 文を整理 (cli の AddArgs/EditArgs/RemoveArgs / input DTO /
  shikomi_core RecordKind/RecordId/RecordLabel/SecretString / time / VaultRepository)
- print_stdout / eprint_stderr に `#[doc(hidden)] pub` 付与 (cross-module 参照)

500 行ルール: lib.rs 338 / record_runners.rs 416、両方 500 行以下クリア。

残作業 (本 commit 範囲外):
- pub fn 自由関数 → 構造体メソッド (ペテルギウス指摘) は accessibility module
  全体を struct + impl に再設計する規模で、別 PR で本格対応する方針
